Regenerated receivers.conf file from lirc-0.7.4
The receivers.conf file was regenerated using the lirc-receiver-list script
and the lirc-0.7.4 sources. All remotes that are not autodetected have been
added to the overrides.conf file to be still available.
Most of the special USB overrides in overrides.conf are now obsolete because
the autodetect script is clever enough to guess the names correctly.